requirements.txt
pandas
Pillow
tqdm
numpy
scikit-learn
scipy
opencv-python-headless
scikit-image